# feat: Optimize pre-commit hook for monorepo performance